# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keeppix API",
    "description": "API di Keeppix. Contratto congelato: solo aggiunte entro /api/v1.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/map/clusters": {
      "get": {
        "tags": [
          "map"
        ],
        "summary": "Get map clusters",
        "operationId": "map_clusters",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "description": "west,south,east,north in gradi WGS84",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zoom",
            "in": "query",
            "description": "Livello di zoom, 0..=30",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "description": "library, album, folder o search",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope_id",
            "in": "query",
            "description": "UUID dello scope",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Celle aggregate o punti singoli",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MapClusterView"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parametri non validi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Scope non visibile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Database non disponibile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/map/tiles/{region}/{z}/{x}/{y}": {
      "get": {
        "tags": [
          "map"
        ],
        "summary": "Serve a map tile from the archive",
        "description": "# Errors\n`401` senza sessione; `404` se la regione è assente o è stata cancellata;\n`416` per un range non valido.",
        "operationId": "map_tile_archive",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "description": "Id della regione locale",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "z",
            "in": "path",
            "description": "Zoom",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "x",
            "in": "path",
            "description": "Coordinata tile X",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "y",
            "in": "path",
            "description": "Coordinata tile Y",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archivio PMTiles"
          },
          "206": {
            "description": "Byte range PMTiles"
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Regione non disponibile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "416": {
            "description": "Range non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/places/reverse": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Reverse-geocode a place",
        "operationId": "places_reverse",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "Latitudine WGS84, -90..=90",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "description": "Longitudine WGS84, -180..=180",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Località o area amministrativa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceView"
                }
              }
            }
          },
          "400": {
            "description": "Coordinate non valide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Nessuna località entro soglia",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/places/suggest": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Suggest places",
        "operationId": "places_suggest",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Nome, almeno due caratteri",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "near_user",
            "in": "query",
            "description": "Favorisce i luoghi vicini alle ultime 50 assegnazioni",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fino a dieci località",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceView"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Query troppo corta o non valida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MapClusterView": {
        "type": "object",
        "required": [
          "lat",
          "lon",
          "count",
          "cover_asset_id",
          "clustered"
        ],
        "properties": {
          "clustered": {
            "type": "boolean"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "cover_asset_id": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "PlaceView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "ascii_name",
          "lat",
          "lon",
          "population"
        ],
        "properties": {
          "admin1": {
            "type": [
              "string",
              "null"
            ]
          },
          "admin2": {
            "type": [
              "string",
              "null"
            ]
          },
          "ascii_name": {
            "type": "string"
          },
          "country_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "population": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "Errore in formato RFC 9457. Il campo `type` è un codice stabile su cui i\nclient possono ramificare; `title` è in inglese e serve al debug, non\nall'utente finale — la traduzione avviene nel frontend.\n\nÈ anche uno schema del documento `OpenAPI`: è la forma su cui il client\nmobile ramifica, quindi va descritta, non solo servita.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "example": "keeppix/unauthenticated"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-kpx_session",
        "description": "Cookie di sessione emesso da POST /api/v1/setup e POST /api/v1/auth/login."
      }
    }
  },
...
}